requirements.txt
requests
pydub
python-telegram-bot
google-auth
google-api-python-client
